# packages/openapi/v1/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/permissions": {
      "get": {
        "description": "Returns list of all available permissions, optionally filtered by domain.",
        "operationId": "PermissionController_findAll",
        "parameters": [
          {
            "name": "domain",
            "required": false,
            "in": "query",
            "description": "Filter by domain (e.g. identity, workspace, engineering)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permissions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all permissions",
        "tags": [
          "permissions"
        ]
      },
      "post": {
        "description": "Creates a new permission with domain.action slug format.",
        "operationId": "PermissionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permission created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Permission with this slug already exists"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new permission",
        "tags": [
          "permissions"
        ]
      }
    },
    "/permissions/{id}": {
      "get": {
        "description": "Returns detailed information about a specific permission.",
        "operationId": "PermissionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Permission UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Permission not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get permission by ID",
        "tags": [
          "permissions"
        ]
      },
      "delete": {
        "description": "Permanently deletes a permission and removes it from all roles.",
        "operationId": "PermissionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Permission UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Permission deleted successfully"
          },
          "404": {
            "description": "Permission not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete permission",
        "tags": [
          "permissions"
        ]
      }
    },
...
    "/ai/agents": {
      "get": {
        "operationId": "AiController_getAgents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Agents retrieved"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List active AI agents",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/conversations": {
      "get": {
        "operationId": "AiController_listConversations",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversations retrieved"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List conversations in workspace",
        "tags": [
          "ai"
        ]
      },
      "post": {
        "operationId": "AiController_createConversation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conversation created"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new conversation",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/conversations/{id}": {
      "get": {
        "operationId": "AiController_getConversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Conversation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation retrieved"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get conversation with messages",
        "tags": [
          "ai"
        ]
      },
      "delete": {
        "operationId": "AiController_deleteConversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Conversation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete conversation",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/conversations/{id}/messages": {
      "post": {
        "operationId": "AiController_sendMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Conversation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI response returned"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Send a message and get AI response",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/usage": {
      "get": {
        "operationId": "AiController_getUsage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Usage stats retrieved"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "AI usage stats for current month",
        "tags": [
          "ai"
        ]
      }
    },
...
  },
  "info": {
    "title": "Xennic Platform API",
    "description": "Xennic Engineering Platform API Documentation",
    "version": "1.0.0",
    "contact": {
      "name": "Xennic Team",
      "url": "https://xennic.com",
      "email": "support@xennic.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://xennic.com/terms"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "PermissionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Permission unique identifier (UUIDv7)"
          },
          "name": {
            "type": "string",
            "description": "Permission display name"
          },
          "slug": {
            "type": "string",
            "description": "Unique permission slug (domain.action)"
          },
          "description": {
            "type": "object",
            "description": "Permission description",
            "nullable": true
          },
          "domain": {
            "type": "string",
            "description": "Domain name"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "domain",
          "createdAt"
        ]
      },
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Read Users",
            "description": "Permission display name"
          },
          "slug": {
            "type": "string",
            "example": "users.read",
            "description": "Unique permission slug in format: domain.action"
          },
          "domain": {
            "type": "string",
            "example": "identity",
            "description": "Domain name: identity | workspace | projects | engineering | ai | marketplace | storage | api | admin"
          },
          "description": {
            "type": "string",
            "example": "Can read user information",
            "description": "Permission description"
          }
        },
        "required": [
          "name",
          "slug",
          "domain"
        ]
      },
...
      "CreateConversationDto": {
        "type": "object",
        "properties": {
          "agentSlug": {
            "type": "string",
            "example": "electrical-engineer"
          },
          "title": {
            "type": "string",
            "example": "بررسی THD سیستم"
          }
        },
        "required": [
          "agentSlug"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "چطور THD جریان را کاهش دهم؟"
          }
        },
        "required": [
          "content"
        ]
      }
    }
  }
}